# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API App - Clinica",
        "version": "1.0"
    },
    "paths": {
...
        "/api/show/actividades-fisicas": {
            "get": {
                "tags": [
                    "ActividadFisica"
                ],
                "summary": "Mostrando las actividaddes fisicas disponibles",
                "operationId": "showActividadesFisicas",
                "responses": {
                    "200": {
                        "description": "Showing aviable Actividades Fisicas"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/actividad-fisica/{actividadFisica}": {
            "get": {
                "tags": [
                    "ActividadFisica"
                ],
                "summary": "Mostrando la actividad fisica",
                "operationId": "showActividadFisica",
                "parameters": [
                    {
                        "name": "actividadFisica",
                        "in": "path",
                        "description": "Id of the Actividad Fisica",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing specific Actividad Fisica"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/create/actividad-fisica": {
            "post": {
                "tags": [
                    "ActividadFisica"
                ],
                "summary": "Crear Actividad Fisica",
                "operationId": "createActividadFisica",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad Fisica created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/update/actividad-fisica/{actividadFisica}": {
            "post": {
                "tags": [
                    "ActividadFisica"
                ],
                "summary": "Actualizar Actividad Fisica",
                "operationId": "updateActividadFisica",
                "parameters": [
                    {
                        "name": "actividadFisica",
                        "in": "path",
                        "description": "Id of the Actividad Fisica",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad Fisica updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/delete/actividad-fisica/{actividadFisica}": {
            "get": {
                "tags": [
                    "ActividadFisica"
                ],
                "summary": "Eliminando la actividad fisica",
                "operationId": "deleteActividadFisica",
                "parameters": [
                    {
                        "name": "actividadFisica",
                        "in": "path",
                        "description": "Id of the Actividad Fisica",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad Fisica deleted successfully"
                    },
                    "403": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}